% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLB_rqss.R
\name{BLB_rqss}
\alias{BLB_rqss}
\alias{FitPlot_1dim}
\title{Bag of Little Bootstraps.}
\usage{
BLB_rqss(cores = NULL, data, alpha = 0.05, tau = 0.25, lambda = 2,
  D = 100, b = ceiling(nrow(data)^0.6), s = 15, r = 100,
  Search = FALSE)

FitPlot_1dim(data, result, xlab = "x", ylab = "y")
}
\arguments{
\item{cores}{The number of cores to use for parallel execution. If not
specified, the number of cores is set to the value of
\code{options("cores")}, if specified, or to one-half the number of cores
detected by the \code{\link{parallel}} package.}

\item{data}{A 2-dim optional data frame (or object coercible by
\code{\link{as.data.frame}} to a data frame) containing the variables in
the model. The column names should be specified in a way that “x” is for
the predictor, and “y” for the response.}

\item{alpha}{The confidence level required. The default is 0.05.}

\item{tau}{A specific quantile to be estimated. Must be a number between 0
and 1.}

\item{lambda}{The smoothing parameter governing the tradeoff between fidelity
and the penalty component for the triogram term. If \code{Search=TRUE},
there is no need for users to specify a value.}

\item{D}{A number that specifies for how many x values you want to compute
confidence intervals; the confidence band is made of pointwise intervals at
varius x values. If specified, it will look at \eqn{D} equidistant points.}

\item{b}{The subsample size in the BLB algorithm. Kleiner et al. suggest that
the size should be around \eqn{n^0.6}, where \eqn{n} is the data size.}

\item{s}{The number of subsamples used in BLB algorithm. Jordan et al.
suggest that \eqn{s} should be 10~20.}

\item{r}{The number of bootstrap iterations (samples with with replacement).
\eqn{r=100} is suggested.}

\item{Search}{If \code{TRUE} (which is recommended), then the function will
first search for an optimum smoothing parameter \eqn{\lambda}.}

\item{xlab, ylab}{Titles for x and y axis. See \code{\link{title}}.}
}
\value{
A list with three parts:

  1. \code{x0} and \code{CI_average}, where \code{x0} contains the x values
  at which the confidence intervals are evaluated, and \code{CI_average} is
  2-dim matrix which contains the corresponding lower and upper bounds.

  2. \code{lambda}, which is the optimum smoothing parameter selected by
  \code{BLB_rqss}. If it is done automatically, the function also returns
  \code{Lambda} and \code{Fid}, which respectively stand for a vector of lambda
  values and their corresponding cross-validation MCV values.
}
\description{
\code{BLB_rqss}  Bag of Litle Bootstraps (BLB) for data sets with 1-dim
covariate. Used to generate  confidence bands for a quantile smoothing
splines fitted with the rqss function from package \code{\link{quantreg}}.

\code{FitPlot_1dim} is a function to plot the results.
}
\details{
This function is based on the "Bag of Little Bootstraps" (BLB) method by
Kleiner et al.(2012). It performs parallelization to speed up the
calculation.
}
\examples{
data(one)
result<-BLB_rqss(cores=7, data=one, alpha=0.05, tau=0.5, Search=TRUE)

plot<-FitPlot_1dim(data=one, result=result, xlab='x',ylab='y')
plot

}
\references{
Micheal, I. J et al.(2012). \eqn{A Scalable Bootstrap for Massive
  Data}.
}
\seealso{
\code{\link{BLB_rqss_2dim}} for BLB with 2-dim covariate dataset.
}
