% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLB_rqss_2dim.R
\name{BLB_rqss_2dim}
\alias{BLB_rqss_2dim}
\alias{FitPlot_2dim}
\title{Bag of Little Bootstraps.}
\usage{
BLB_rqss_2dim(cores = NULL, data, alpha = 0.05, tau = 0.25, lambda = 2,
  D = 50, b = ceiling(nrow(data)^0.6), s = 15, r = 100,
  Search = FALSE)

FitPlot_2dim(data, result, Range = c(0.002, 0.998), xlab = "log(avtp_med)",
  ylab = "log(niqr_med)", zlab = "log(playdelay_med)")
}
\arguments{
\item{cores}{The number of cores to use for parallel execution. If not
specified, the number of cores is set to the value of
\code{options("cores")}, if specified, or to one-half the number of cores
detected by the \code{\link{parallel}} package.}

\item{data}{A 3-dim optional data frame (or object coercible by
\code{\link{as.data.frame}} to a data frame) containing the variables in the
model. The column names should be specified in a way that “x1” and "x2" are
for the two predictors, and “y” for the response.}

\item{alpha}{The confidence level required. The default is 0.05.}

\item{tau}{A specific quantile to be estimated. Must be a number between 0 and
1.}

\item{lambda}{The smoothing parameter governing the tradeoff between fidelity
and the penalty component for the triogram term. If \code{Search=TRUE},
there is no need for users to specify a value.}

\item{D}{A number that specifies for how many x values you want to compute
confidence intervals; the confidence surface is made of pointwise intervals
at varius x values. If specified, it will compute intervals on an
equidistant \eqn{D×D} grid.}

\item{b}{The subsample size in the BLB algorithm. Kleiner et al. suggest that
the size should be around \eqn{n^0.6}, where \eqn{n} is the data size.}

\item{s}{The number of subsamples used in BLB algorithm. Kleiner et al.
suggest that \eqn{s} should be 10~20.}

\item{r}{The number of bootstrap iterations (samples with with replacement).
\eqn{r=100} is suggested.}

\item{Search}{If \code{TRUE} (which is recommended), then the function will
first search for an optimum smoothing parameter \eqn{\lambda}.}

\item{xlab, ylab, zlab}{Titles for the axises. See \code{\link{title}}.}
}
\value{
A list with three parts:

 1. \code{x0} and \code{CI_average}, where \code{x0} contains the x values at
 which the confidence intervals are evaluated, and \code{CI_average} is 2-dim
 matrix which contains the corresponding lower and upper bounds.

 2. \code{lambda}, which is the optimum smoothing parameter selected by
 \code{BLB_rqss_2dim}. If it is done automatically, the function also returns
 \code{Lambda} and \code{Fid}, which respectively stand for a vector of lambda
 values and their corresponding cross-validation MCV values.
}
\description{
\code{BLB_rqss_2dim}  Bag of Litle Bootstraps (BLB) for data sets with 2-dim
covariate. Used to generate confidence bands for a quantile smoothing splines
fitted with the rqss function from package \code{\link{quantreg}}.

\code{FitPlot_2dim} is a function to plot the results.
}
\details{
This function is based on the "Bag of Little Bootstraps" (BLB) method by
Kleiner et al.(2012). It performs parallelization to speed up the calculation.
}
\examples{
data(two)
result<-BLB_rqss_2dim(cores=7, data=two, alpha=0.05, tau=0.5, Search=TRUE)

#1. 3D plot
plot<-FitPlot_2dim(data=two, result=result, xlab='x', ylab='y', zlab='z')
plot

#2. Contour plot
#Plot the confidence bands
library(akima)
valid<-!is.na(result$CI_average[,1])
X<-result$x0$x1[valid];Y<-result$x0$x2[valid]
lower<-result$CI_average[valid,1]; upper<-result$CI_average[valid,2]
x1range<-range(X);x2range<-range(Y)
akima1<-interp(X,Y,lower,xo=seq(x1range[1],x1range[2],length=200),yo=seq(x2range[1],x2range[2],length=200))
akima2<-interp(X,Y,upper,xo=seq(x1range[1],x1range[2],length=200),yo=seq(x2range[1],x2range[2],length=200))

plot(X,Y,main="Prediction Contour Plot",type="n",xlab="x1",ylab="x2")
image(akima1,add=TRUE,col = terrain.colors(12))  ###YOU CAN CHANGE THE NUMBER OF COLOR LEVELS!!!!
contour(akima1,add=TRUE,nlevels=12)


}
\references{
Kleiner, I. J et al. JRSS B, 2012. \eqn{A Scalable Bootstrap for
 Massive Data}.

Roger Koenker.(2005). \eqn{Quantile regression}. Cambridge
 university press.
}
\seealso{
\code{\link{BLB_rqss}} for BLB with 1-dim covariate dataset.
}
