% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\docType{data}
\name{control}
\alias{control}
\title{One-covariate control dataset for comparative experiment}
\format{A data frame with 3200 rows and 2 variables: \describe{
  \item{x}{covariate}
  \item{y}{response} ... }}
\usage{
control
}
\description{
A dataset containing the covariate 'x' and the response 'y' for 3,200
observations.
}
\keyword{datasets}
