% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldsGenerator.R
\name{foldsGenerator}
\alias{foldsGenerator}
\title{Fold Generator for Cross-Validation.}
\usage{
foldsGenerator(sim.data, nfolds = 10, n = NULL)
}
\arguments{
\item{sim.data}{A data frame (or object coercible by
\code{\link{as.data.frame}} to a data frame) on which the cross validation
is performed.}

\item{nfolds}{The number of folds for the cross validation.}

\item{n}{The number of rows if the data frame not provided.}
}
\value{
A vector of random numbers which vary from \eqn{1} to \eqn{nfolds}.
}
\description{
\code{foldsGenerator} is used to generate folds for a specific data set when
performing crossvalidation.
}
\details{
This function first identifies the number of observations in the data set.
With the pre-specified number of folds \code{nfolds}, the function generates a
sequence of randoms numbers which vary from \eqn{1} to \eqn{nfolds}. Each
number indicates which fold an observation belongs to.
}
\examples{
data(one)

folds<-foldsGenerator(one,nfolds=10)
}
