% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifoldCV.R
\name{multifoldCV}
\alias{multifoldCV}
\alias{multifoldCV_2dim}
\alias{rho}
\title{MCV Criterion for the Selection of \eqn{\lambda}.}
\usage{
multifoldCV(cores = NULL, lambda, tau, sim.data, cv.idx)

multifoldCV_2dim(cores = NULL, lambda, tau, sim.data, cv.idx)

rho(tau, x)
}
\arguments{
\item{cores}{The number of cores to use for parallel execution. If not
specified, the number of cores is set to the value of
\code{options("cores")}, if specified, or to one-half the number of cores
detected by the \code{\link{parallel}} package.}

\item{lambda}{The smoothing parameter governing the tradeoff between fidelity
and the penalty component for the triogram term.}

\item{tau}{A specific quantile to be estimated. Must be a number between 0 and
1.}

\item{sim.data}{A data frame (or object coercible by
\code{\link{as.data.frame}} to a data frame) for which we are selecting the
smoothing parameter.}

\item{cv.idx}{The folds (the divided data) randomly generated by
\code{\link{foldsGenerator}}.}
}
\value{
The MCV criterion is returned.
}
\description{
\code{multifoldCV} calculates the multi-fold crossvalidation (MCV) criterion
for choosing the smoothness parameter in quantile smoothing splines. The MCV
criterion was proposed by Reiss et al. in 2012. Reiss et al. and our own
simulations demonstrate this is criterion is suitable as a reliable automatic
 procedure to choose the optimal smoothing parameter.

\code{multifoldCV_2dim} calculates the same criterion for 2-dim covariate data sets.

\code{rho} is a function to plot the results.
}
\details{
In general, the optimal degree of smoothness will depend on the quantile of
interest. Given a value \eqn{\tau\in [0,1]}, nonparametric quantile regression
seeks to estimate the conditional quantile function \eqn{g(x)} minimizing:
\deqn{\sum_{i=1}^n \rho_{\tau}[y_i-g(x_i)]+\lambda J(g)}
where J(g) is the total variation of g(x).
Multifold CV (Reiss, 2012) can greatly reduce the computational burden. Here
we divide the \eqn{n} observations into \eqn{V_1,\cdots,V_k} of
(approximately) equal size, and define the criterion:
\deqn{MCV=\frac{1}{n}\sum_{j=1}^k\sum_{i\in V_j}
\rho_{\tau}[y_i-\hat{g}_{\lambda}^{[-V_j]}(x_i)]} where
\eqn{\hat{g}_{\lambda}^{[-V_j]}} is the function estimate based on all
observations except those in fold \eqn{V_j}. Small values of \eqn{k}
produce downward-biased estimates of prediction error, whereas larger values
produce more variable results and impose a higher computational burden;
\eqn{k=5 or 10} is often recommended as a compromise.
}
\examples{
data(one)
folds<-foldsGenerator(sim.data=one,nfolds=10)
multifoldCV(cores=3,lambda=2,tau=0.5,sim.data=one,cv.idx=folds)

data(two)
folds<-foldsGenerator(sim.data=two,nfolds=10)
multifoldCV_2dim(cores=3,lambda=2,tau=0.5,sim.data=two,cv.idx=folds)

}
\references{
Reiss et al. (2012). \eqn{Smoothness Selection for penalized Quantile Regression Splines}.
}
\seealso{
\code{\link{foldsGenerator}}
}
