% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.R
\docType{data}
\name{one}
\alias{one}
\title{One-covariate dataset for Testing}
\format{A data frame with 2999 rows and 2 variables:
\describe{
  \item{x}{covariate}
  \item{y}{response}
  ...
}}
\usage{
one
}
\description{
A dataset containing the covariate 'x' and the response 'y' for almost 3,000
observations.
}
\keyword{datasets}
