% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.data.R
\docType{data}
\name{sim.data}
\alias{sim.data}
\title{A data set which is used in the paper.}
\format{A data frame with 5,000 rows and 2 variables:
\describe{
  \item{x}{covariate}
  \item{y}{response}
  ...
}}
\usage{
sim.data
}
\description{
A data set which is used to demonstrate that the conventional SIC and other 
metrics would fail in some cases.
}
\keyword{datasets}
