% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment.R
\docType{data}
\name{treatment}
\alias{treatment}
\title{One-covariate treatment dataset for comparative experiment}
\format{A data frame with 3200 rows and 2 variables: \describe{
  \item{x}{covariate}
  \item{y}{response} ... }}
\usage{
treatment
}
\description{
A dataset containing the covariate 'x' and the response 'y' for 3,200
observations.
}
\keyword{datasets}
