% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_2dim.R
\docType{data}
\name{treatment_2dim}
\alias{treatment_2dim}
\title{Two-covariate treatment dataset for comparative experiment}
\format{A data frame with 3200 rows and 2 variables: \describe{
  \item{x1}{covariate}
  \item{x2}{covariate}
  \item{y}{response} ... }}
\usage{
treatment_2dim
}
\description{
A dataset containing the covariates 'x1', 'x2', and the response 'y' for
3,200 observations.
}
\keyword{datasets}
