% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.R
\docType{data}
\name{two}
\alias{two}
\title{Two-covariate dataset for Testing}
\format{A data frame with 2999 rows and 2 variables:
\describe{
  \item{x1}{first covariate}
  \item{x2}{second covariate}
  \item{y}{response}
  ...
}}
\usage{
two
}
\description{
A dataset containing the covariates 'x1', "x2" and the response 'y' for
almost 3,000 observations.
}
\keyword{datasets}
