% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vignette_data.R
\docType{data}
\name{Vignette_data}
\alias{Vignette_data}
\title{A data set for complete demonstration of the package}
\format{A data frame with 50,000 rows and 4 variables:
\describe{
  \item{cell}{Cell indicator}
  \item{x1}{covariate}
  \item{y2}{covariate}
  \item{z}{response}
  ...
}}
\usage{
Vignette_data
}
\description{
A data set which is a part of the Test 7928. It contains the contains 
two covariates and one response for 50,000 observations.
}
\keyword{datasets}
